refactoring of BuildDefinitionsList
